Part 1 Modeling attempts using features
nb_02: Now we did all the neede EDA for the data collected
nb_03: Now we finished the baseline model for the data
This model uses a random forest method called XGBOOST
Although the results are very good, but this model falls short of being realistic since it assumes we have the data of the features at the time of prediction.
nb_04: Now we have completed a proper forecasting of the energy prices using recursive prediction with XGBOOST
The shortcomings of this method is that it dos not use the provided features and only relys on the history of the energy prices
nb_05: This notebook was to try and forecast the features and use this forecast to predict the energy prices accordingly
However, the shortcoming of this method is that it relies on the actual data to predict the future data. Meaning it can predict the solar energy production for the next hour because it has the current hour. however, it predicts the solar energy production of the next next hour based on the test set value of the next hour which in real life is not available
nb_06 solar: An attempt to properly predict solar energy outcome which could not yield any proper useful results
nb_06 all: Trying to forecast the features 1 day, 1 week, 1 month.....12 months in order to use the data as input to predict the energy price of a certain day. But the results end up being periodical and do not represent reallife data
nb_07: Tried integrating the method of the 06 notebook but only for a week to see if it is good enough to predict the prices but the results ended up being very bad especially compared to the simple model in notebook 04
nb_08: Tried a very computationally expensive model to see if it works better. The input x was all the features and their history for 168 hours in the past. The output was 168 data points representing the price of energy at time t+1 to t+168. Results show that the method in notebook 04 is still better
nb_09: Tried collecting news data for the next part of the project but no articles were found.